Update README.md to include THL link (#262)
* Update README.md to add THL

Based on the hundreds of thousands of #angular2 tweets already processed, noise is a little over 90% of the conversation. The platform filters and ranks the most popular shared content in realtime. Angular2’s zeitgeist, you might say. It's been running for over a year, monitoring half a billion tweets a day, and will always be free to use. No ads. No BS.

* Update README.md to include The Herd Locker

Based on the hundreds of thousands of #angular2 tweets already processed, noise is a little over 90% of the conversation. The platform filters and ranks the most popular shared content in realtime. Angular2’s zeitgeist, you might say. It's been running for over a year, monitoring half a billion tweets a day, and will always be free to use. No ads. No BS.

* Update README.md to include THL (capitalised)

Based on the hundreds of thousands of #angular2 tweets already processed, noise is a little over 90% of the conversation. The platform filters and ranks the most popular shared content in realtime. Angular2’s zeitgeist, you might say. It's been running for over a year, monitoring half a billion tweets a day, and will always be free to use. No ads. No BS.